requirements.txt
absl-py==0.9.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
boto3==1.14.33
botocore==1.17.33
Brotli==1.0.7
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.13.1
chardet==3.0.4
Django==3.0.7
django-appconf==1.0.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.0.1
django-crispy-forms==1.9.2
django-extensions==3.0.4
django-filter==2.3.0
django-select2==7.4.2
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==1.15.0
docutils==0.15.2
importlib-metadata==1.6.1
incremental==17.5.0
inflection==0.5.0
ipykernel==5.3.0
ipython @ file:///opt/concourse/worker/volumes/live/02237e64-34e3-452e-5c63-14b73548ce0e/volume/ipython_1592348474440/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itemadapter==0.1.0
itsdangerous==1.1.0
jedi @ file:///opt/concourse/worker/volumes/live/114a62c0-dd60-4a40-6382-85abb31b5f6e/volume/jedi_1592841953086/work
jmespath==0.10.0
joblib==0.15.1
json-excel-converter==1.3.0
json5==0.9.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.0.0
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
jupyterthemes==0.20.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lesscpy==0.14.0
libarchive-c==2.9
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mongoengine==0.20.0
more-itertools==8.4.0
mrjob==0.7.3
multidict==4.7.6
mysql-connector-python==8.0.20
mysqlclient==2.0.1
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.3
numpy==1.18.1
olefile==0.46
packaging==20.4
pandas==1.0.5
pandas-datareader==0.8.1
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==2.2.2
pkginfo==1.5.0.1
plotly==4.8.2
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.12.1
proxy-requests==0.5.1
psutil==5.7.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
PyDispatcher==2.0.5
pydotplus==2.0.2
Pygments==2.4.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.15.4
PySocks==1.7.1
pyspark==3.0.0
pytest==3.7.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.7.5
QtPy==1.9.0
Quandl==3.5.1
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
retrying==1.3.3
rsa==4.6
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
Scrapy==2.2.1
seaborn==0.10.1
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
slackclient==1.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlparse==0.3.1
statsmodels==0.11.1
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
terminado==0.8.3
traitlets==4.3.3
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.27
w3lib==1.22.0
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.14.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
XlsxWriter==1.3.3
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
zope.interface==5.1.0
